github-oidc-bootstrap/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "github-oidc-bootstrap"
version = "0.1.0"
description = "GitHub OIDC provider and IAM role for Infiquetra Organizations CDK deployments"
readme = "README.md"
authors = [
    {name = "Infiquetra", email = "admin@infiquetra.com"}
]
license = {text = "MIT"}
requires-python = ">=3.11"
dependencies = [
    "aws-cdk-lib>=2.150.0",
    "constructs>=10.3.0",
    "python-dotenv>=1.0.0",
    "boto3>=1.34.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "boto3-stubs[essential]>=1.34.0",
    "cfn-lint>=0.85.0",
    "bandit>=1.7.5",
    "mypy>=1.8.0",
    "ruff>=0.1.0",
    "coverage[toml]>=7.0.0",
]

[tool.ruff]
target-version = "py311"
line-length = 88

[tool.ruff.lint]
extend-select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
    "S",  # bandit security
]
ignore = [
    "E501",  # line too long, handled by black
    "S101",  # use of assert
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.coverage.run]
source = ["github_oidc_bootstrap"]
omit = [
    "tests/*",
    "*/tests/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]